Add mocked test for retrying publishing
