+++ cli/src/pixl_cli/_message_processing.py
@@ -50,7 +50,7 @@ def retry_until_export_count_is_unchanged(
     )
     for i in range(1, num_retries + 1):
         _wait_for_queues_to_empty(queues_to_populate)
-        logger.info("Waiting 5 minutes for new extracts to be found")
+        logger.info("Waiting {} for new extracts to be found", wait_to_display)
         for _ in tqdm.tqdm(
             range(total_wait_seconds), desc="Waiting for series to be fully processed"
         ):
